optimisation de la classe Log + implémentation de la classe Processor + optimisation de la classe Processor + ajout de frames + amélioration de la reflection

+++ messages/connection/HelloConnectMessage.java
@@ -27,7 +27,7 @@ public class HelloConnectMessage extends Message {
 		int len = this.key.length;
 		buffer.writeVarInt(len);
 		for(int i = 0; i < len; ++i)
-			buffer.writeByte((byte) this.key[i]);
+			buffer.writeByte(this.key[i]);
 		
 		completeInfos(buffer);
 	}
